% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_greeks.R
\name{bsm_delta}
\alias{bsm_delta}
\title{The delta in the Black-Scholes-Merton corresponds to the first derivative of
the BSM function with respect to x. It practically refers to the amount of
stock to keep in order to hedge an call/put option}
\usage{
bsm_delta(initial_stock_price = 50, time_to_maturity = 4, seed = 1,
  scale = 100, sigma = 1, alpha = 0, strike = initial_stock_price,
  riskless_rate = 0.03)
}
\arguments{
\item{strike}{Strike price defined for the stock option}

\item{riskless_rate}{Riskless rate}
}
\value{

}
\description{
Delta
}
