% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_greeks.R
\name{bsm_gamma}
\alias{bsm_gamma}
\title{The gamma in the Black-Scholes-Merton corresponds to the second derivative of
the BSM function with respect to x^2.}
\usage{
bsm_gamma(s, sigma = 1, strike = initial_stock_price,
  riskless_rate = 0.03)
}
\value{

}
\description{
gamma
}
