% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_greeks.R
\name{bsm_theta}
\alias{bsm_theta}
\title{The theta in the Black-Scholes-Merton corresponds to the first derivative of
the BSM function with respect to time(t). It can be thought as the
instantaneous change in BSM price when time move forward}
\usage{
bsm_theta(s, sigma = 1, strike = initial_stock_price,
  riskless_rate = 0.03)
}
\value{

}
\description{
(!) due to the zero division as t -> T, the final value when t = T is not a
number
}
\details{
Theta
}
