% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_hedging.R
\name{delta_hedging}
\alias{delta_hedging}
\title{This function compute the delta hedging for time series following
Geometric Brownian Motion. i.e. log-return are normally distributed.}
\usage{
delta_hedging(time_to_maturity, S, strike, sigma = 0.2, alpha = 0.05,
  r = 0.03, scale = 365, full = F)
}
\arguments{
\item{time_to_maturity}{Time up to maturity in year}

\item{S}{Initial value of the Stock}

\item{strike}{Strike price of the option.}

\item{sigma}{Volatility rate of the simulated stock time series}

\item{alpha}{Drift rate of the simulated stock time series}

\item{r}{Riskless interest rate - 0.03 if no value provided.}

\item{scale}{Year divided factor}

\item{full}{By defaut the function only return the delta hedging results}
}
\value{

}
\description{
Delta hedging of Black-Scholes_Merton
}
\examples{


}
