% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_greeks.R
\name{gamma}
\alias{gamma}
\title{The gamma in the Black-Scholes-Merton corresponds to the second derivative of
the BSM function with respect to x^2.}
\usage{
gamma(initial_stock_price = 50, time_to_maturity = 4, seed = 1,
  scale = 100, sigma = 1, alpha = 0, strike = initial_stock_price,
  riskless_rate = 0.03)
}
\arguments{
\item{initial_stock_price}{Stock price at initial step}

\item{time_to_maturity}{Time upon which measurement is done}

\item{seed}{Control randmoness}

\item{scale}{Divided part of a unit. For instance scale of 100 will divide
4 units of time in 100 parts each, that brings to a total time of
400 periods of time measurement.}

\item{sigma}{Volatility of the stock}

\item{alpha}{Mean rate of return of the stock}

\item{strike}{Strike price defined for the stock option}

\item{riskless_rate}{Riskless rate}
}
\value{

}
\description{
gamma
}
