% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstock.R
\name{mjd_ts}
\alias{mjd_ts}
\title{Jump Diffusion Model}
\usage{
mjd_ts(initial_stock_price = 50, time_to_maturity = 5, seed = 1,
  scale = 365, sigma = 0.2, alpha = 0, lambda = 7,
  jumps_intensity_parameters = list(mean = 0, sd = 0.2))
}
\value{
A data.frame containing 2 variables: stock_price_path and the
 according time_periods.
}
\description{
Stock price model with jumps.
}
\author{
Anthony Tedde
}
