% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstock.R
\name{sstock}
\alias{sstock}
\title{Stock ticker simulation}
\usage{
sstock(initial_stock_price = 50, time_to_maturity = 4, seed = 1,
  scale = 100, sigma = 1, alpha = 0)
}
\description{
\code{sstock} outputs price path for a fully random stock evolution
}
