% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstock.R
\name{sstock_ito}
\alias{sstock_ito}
\title{Stock Price Approximation using Itô}
\usage{
sstock_ito(initial_stock_price = 50, time_to_maturity = 4, seed = 1,
  scale = 100, sigma = 1, alpha = 0)
}
\value{
A data.frame containing 2 variables: stock_price_path and the
 according time_periods.
}
\description{
This function approximates the path of a stock price driven by a unique
Brownian Motion, using the Itô-Doeblin formula for Itô-processes.
}
\author{
Anthony Tedde
}
