% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate_model.R
\name{vasicek_ir}
\alias{vasicek_ir}
\title{Title}
\usage{
vasicek_ir(initial_interest_rate = 0.1, time_to_maturity = 4, scale = 100,
  seed = 1, alpha = 0, beta = 1, sigma = 1)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
