% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_scholes_merton.R
\name{BSM}
\alias{BSM}
\title{Black-Scholes_Merton equation}
\usage{
BSM(delta_time, current_stock_price, strike, interest_rate, sigma)
}
\arguments{
\item{sigma}{}
}
\description{
Black-Scholes_Merton equation
}
